requirements.txt
streamlit==1.14.0
transformers==4.24.0
newspaper3k==0.2.8
fpdf==1.7.2
